+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## [0.0.2] - First Release with drag 
+* Change naming for custom animation controller
 ## [0.0.1] - First Release with drag helper
 
 * Create Drag helper